Move `Package` to package.rs
