+++ crates/oak_db/src/lib.rs
@@ -33,11 +33,11 @@ pub use identifier::NamespacePart;
 pub use inputs::LibraryRoots;
 pub use inputs::LiveRoot;
 pub use inputs::OrphanRoot;
-pub use inputs::Package;
 pub use inputs::Root;
 pub use inputs::RootKind;
 pub use inputs::StaleRoot;
 pub use inputs::WorkspaceRoots;
 pub use name::Name;
+pub use package::Package;
 pub use package_resolve::PackageVisibility;
 pub use storage::OakDatabase;
